Server

+++ dht11.h
@@ -2,5 +2,5 @@
 #include "lib.h"
 
 void dht11_init();
-bool dht11_read(unsigned int* temperature, unsigned int* humidity);
+bool dht11_read(int8_t* temperature, uint8_t* humidity);
